# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RRHH API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/legajos": {
            "get": {
                "tags": [
                    "Legajo"
                ],
                "summary": "getLegajoList",
                "description": "Get all Legajos",
                "operationId": "b90528b43ea4f112918999d0c5cd8e8a",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Legajo"
                ],
                "summary": "createLegajo",
                "description": "Create Legajo",
                "operationId": "d1d75fcf6e8be699f3b7fcaf42523e8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/legajos/{id}": {
            "get": {
                "tags": [
                    "Legajo"
                ],
                "summary": "getLegajoItem",
                "description": "Get Legajo",
                "operationId": "5b3f0a954b6e2d552e0beaf7f8824946",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Legajo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Legajo"
                ],
                "summary": "updateLegajo",
                "description": "Update Legajo",
                "operationId": "ef2f9f0d3aaf32700507ae7d4f0b6cd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Legajo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Legajo"
                ],
                "summary": "deleteLegajo",
                "description": "Delete Legajo",
                "operationId": "f62c21089539d3b637c3f1784ab37978",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Legajo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    }
}